feat: updated module structure to use a prelude
