+++ TODO.md
@@ -1,1 +1,2 @@
 Not require explicitely specifying Visibility::Visible by updating to required components
+Migrate to using bevy builtin CosmicFontSystem resource and CosmicBuffer component
