+++ src/lib.rs
@@ -82,217 +82,62 @@
 //! MIT or Apache-2.0
 #![allow(clippy::type_complexity)]
 
+pub mod prelude {
+    // external re-exports
+    pub(crate) use bevy::prelude::*;
+
+    // internal re-exports
+    pub(crate) use crate::buffer::BufferExtras as _;
+    pub(crate) use crate::cosmic_text;
+    pub(crate) use crate::primary::CosmicRenderOutput;
+    pub(crate) use crate::primary::NodeSizeExt as _;
+    pub(crate) use crate::utils::*;
+
+    // public internal re-exports
+    pub use crate::buffer::CosmicBuffer; // todo: migrate to builtin bevy CosmicBuffer
+    pub use crate::cosmic_edit::CosmicFontSystem; // todo: migrate to using builtin bevy cosmic font system
+    pub use crate::cosmic_edit::{
+        CosmicEditBundle, CosmicEditor, CosmicSource, DefaultAttrs, ReadOnly,
+    };
+    pub use crate::focus::FocusedWidget;
+    pub use crate::primary::{CosmicEditPlugin, CosmicFontConfig, CosmicPrimaryCamera};
+    pub use crate::utils::{
+        change_active_editor_sprite, change_active_editor_ui, deselect_editor_on_esc,
+        print_editor_text, ColorExtras as _,
+    };
+    #[doc(no_inline)]
+    pub use bevy::text::cosmic_text::{
+        Color as CosmicColor, Style as FontStyle, Weight as FontWeight,
+    };
+}
+
+pub use bevy::text::cosmic_text;
+
+pub use primary::{CosmicEditPlugin, CosmicFontConfig, CosmicPrimaryCamera, CosmicRenderOutput};
+/// Contains the library global important types you probably want to explore first
+mod primary;
+
+pub use buffer::CosmicBuffer;
 mod buffer;
+pub use cosmic_edit::{
+    CosmicBackgroundColor, CosmicBackgroundImage, CosmicEditor, CosmicFontSystem, CosmicSource,
+    CosmicTextAlign, CosmicWrap, CursorColor, DefaultAttrs, MaxChars, MaxLines, ReadOnly,
+    SelectedTextColor, SelectionColor, XOffset,
+};
 mod cosmic_edit;
+pub use cursor::{HoverCursor, TextHoverIn, TextHoverOut};
 mod cursor;
+pub use events::CosmicTextChanged;
 mod events;
+pub use focus::FocusedWidget;
 mod focus;
+pub use input::InputSet;
 mod input;
+pub use password::Password;
 mod password;
+pub use placeholder::Placeholder;
 mod placeholder;
 mod render;
 mod user_select;
-mod util;
+pub mod utils;
 mod widget;
-
-use std::{path::PathBuf, time::Duration};
-
-use bevy::{prelude::*, utils::tracing::instrument};
-
-pub use buffer::*;
-pub use cosmic_edit::*;
-#[doc(no_inline)]
-pub use cosmic_text::{self, Color as CosmicColor, Style as FontStyle, Weight as FontWeight};
-pub use cursor::*;
-pub use events::*;
-pub use focus::*;
-pub use input::*;
-pub use password::*;
-pub use placeholder::*;
-pub use render::*;
-pub use user_select::*;
-pub use util::*;
-pub use widget::*;
-
-/// Plugin struct that adds systems and initializes resources related to cosmic edit functionality.
-#[derive(Default)]
-pub struct CosmicEditPlugin {
-    pub font_config: CosmicFontConfig,
-}
-
-impl Plugin for CosmicEditPlugin {
-    fn build(&self, app: &mut App) {
-        trace!("Loading cosmic edit plugin");
-        let font_system = create_cosmic_font_system(self.font_config.clone());
-
-        app.add_plugins((
-            BufferPlugin,
-            RenderPlugin,
-            WidgetPlugin,
-            InputPlugin,
-            FocusPlugin,
-            CursorPlugin,
-            PlaceholderPlugin,
-            PasswordPlugin,
-            EventsPlugin,
-            UserSelectPlugin,
-        ))
-        .insert_resource(CosmicFontSystem(font_system));
-
-        app.register_type::<CosmicWrap>()
-            .register_type::<CosmicTextAlign>()
-            .register_type::<XOffset>()
-            .register_type::<CosmicBackgroundImage>()
-            .register_type::<CosmicBackgroundColor>()
-            .register_type::<CursorColor>()
-            .register_type::<SelectionColor>()
-            .register_type::<MaxLines>()
-            .register_type::<MaxChars>()
-            .register_type::<CosmicSource>()
-            .register_type::<HoverCursor>();
-
-        #[cfg(target_arch = "wasm32")]
-        {
-            let (tx, rx) = crossbeam_channel::bounded::<WasmPaste>(1);
-            app.insert_resource(WasmPasteAsyncChannel { tx, rx })
-                .add_systems(Update, poll_wasm_paste);
-        }
-    }
-}
-
-/// Attach to primary camera, and enable the `multicam` feature to use multiple cameras.
-/// Will panic if no Camera's without this component exist and the `multicam` feature is enabled.
-///
-/// A very basic example which doesn't panic:
-/// ```rust
-/// use bevy::prelude::*;
-/// use bevy_cosmic_edit::CosmicPrimaryCamera;
-///
-/// fn main() {
-///   App::new()
-///     .add_plugins((
-///       DefaultPlugins,
-///       bevy_cosmic_edit::CosmicEditPlugin::default(),
-///     ))
-///     .add_systems(Startup, setup)
-///     .run();
-/// }
-///
-/// fn setup(mut commands: Commands) {
-///   commands.spawn((Camera3dBundle::default(), CosmicPrimaryCamera));
-///   commands.spawn(Camera3dBundle {
-///     camera: Camera {
-///       order: 2,
-///       ..default()
-///     },
-///     ..default()
-///   });
-/// }
-/// ```
-#[derive(Component, Debug, Default)]
-pub struct CosmicPrimaryCamera;
-
-/// Resource struct that holds configuration options for cosmic fonts.
-#[derive(Resource, Clone)]
-pub struct CosmicFontConfig {
-    pub fonts_dir_path: Option<PathBuf>,
-    pub font_bytes: Option<Vec<&'static [u8]>>,
-    /// If [false], some characters (esspecially Unicode emojies) might not load properly
-    /// Caution: this can be relatively slow
-    pub load_system_fonts: bool,
-}
-
-impl Default for CosmicFontConfig {
-    fn default() -> Self {
-        let fallback_font = include_bytes!("./font/FiraMono-Regular-subset.ttf");
-        Self {
-            load_system_fonts: true,
-            font_bytes: Some(vec![fallback_font]),
-            fonts_dir_path: None,
-        }
-    }
-}
-
-/// Used to ferry data from a [`CosmicBuffer`] to the receiving entity with [`CosmicSource`]
-///
-/// See [crate::buffer::swap_target_handle]
-#[derive(Component, Default, Debug, Deref, DerefMut)]
-pub struct CosmicRenderOutput(pub Handle<Image>);
-
-trait NodeSizeExt {
-    fn logical_size(&self) -> Vec2;
-}
-
-impl NodeSizeExt for ComputedNode {
-    #[instrument]
-    fn logical_size(&self) -> Vec2 {
-        self.size() * self.inverse_scale_factor()
-    }
-}
-
-fn create_cosmic_font_system(cosmic_font_config: CosmicFontConfig) -> cosmic_text::FontSystem {
-    let locale = sys_locale::get_locale().unwrap_or_else(|| String::from("en-US"));
-    let mut db = cosmic_text::fontdb::Database::new();
-    if let Some(dir_path) = cosmic_font_config.fonts_dir_path.clone() {
-        db.load_fonts_dir(dir_path);
-    }
-    if let Some(custom_font_data) = &cosmic_font_config.font_bytes {
-        for elem in custom_font_data {
-            db.load_font_data(elem.to_vec());
-        }
-    }
-    if cosmic_font_config.load_system_fonts {
-        db.load_system_fonts();
-    }
-    cosmic_text::FontSystem::new_with_locale_and_db(locale, db)
-}
-
-#[cfg(test)]
-mod tests {
-    use bevy::input::keyboard::KeyboardInput;
-
-    use crate::*;
-
-    use self::buffer::CosmicBuffer;
-
-    fn test_spawn_cosmic_edit_system(
-        mut commands: Commands,
-        mut font_system: ResMut<CosmicFontSystem>,
-    ) {
-        let attrs = cosmic_text::Attrs::new();
-        commands.spawn(CosmicEditBundle {
-            buffer: CosmicBuffer::new(&mut font_system, cosmic_text::Metrics::new(20., 20.))
-                .with_rich_text(&mut font_system, vec![("Blah", attrs)], attrs),
-            ..Default::default()
-        });
-    }
-
-    #[test]
-    fn test_spawn_cosmic_edit() {
-        let mut app = App::new();
-        app.add_plugins(TaskPoolPlugin::default());
-        app.add_plugins(AssetPlugin::default());
-        app.insert_resource(CosmicFontSystem(create_cosmic_font_system(
-            CosmicFontConfig::default(),
-        )));
-        app.add_systems(Update, test_spawn_cosmic_edit_system);
-
-        let input = ButtonInput::<KeyCode>::default();
-        app.insert_resource(input);
-        let mouse_input: ButtonInput<MouseButton> = ButtonInput::<MouseButton>::default();
-        app.insert_resource(mouse_input);
-
-        app.add_event::<KeyboardInput>();
-
-        app.update();
-
-        let mut text_nodes_query = app.world_mut().query::<&CosmicBuffer>();
-        for cosmic_editor in text_nodes_query.iter(&app.world()) {
-            insta::assert_debug_snapshot!(cosmic_editor
-                .lines
-                .iter()
-                .map(|line| line.text())
-                .collect::<Vec<_>>());
-        }
-    }
-}
